Split action validation
