+++ CMakeLists.txt
@@ -206,6 +206,7 @@ target_sources(
         src/game_data_world_queries.c
         src/game_data_standard_options.c
         src/game_presentation.c
+        src/game_data_validation_actions.c
         src/game_data_validation_brush_common.c
         src/game_data_validation_app_world.c
         src/game_data_validation_actors.c
